feat(runtime): uplift rate limiter from desktop to dasclaw_runtime::rate_limit
Verbatim port of desktop-client/ironclaw/src/tools/rate_limiter.rs to
crates/dasclaw_runtime/src/rate_limit.rs. The desktop file becomes a
thin re-export shim, preserving all existing call sites:

  - tools/registry.rs
  - tools/wasm/rate_limiter.rs
  - tools/wasm/mod.rs
  - tools/mod.rs
  - worker/job.rs

This is ADR-153 step 2 preparation: the headless agent runtime needs
rate limiting to be available from the runtime crate (not buried in
the desktop binary) before the Agent::builder facade can land in a
follow-up PR.

Cross-cuts: 不涉及 ADR-114 (no new .ironclaw / IRONCLAW_BASE_DIR
literals; pure crate-level move).

Verification:
  cargo check -p dasclaw_runtime --tests       # 0 err 0 warn
  cargo check -p dasclaw --tests               # 0 err 0 warn
  cargo nextest run -p dasclaw_runtime         # 9 tests pass
  cargo clippy --no-deps -p dasclaw_runtime --all-targets -- -D warnings
  cargo fmt --all && python3.12 scripts/check_no_panics.py --base origin/xClaw

已检查 dasclaw_runtime 是否已有 rate_limiter，结论：未实现（grep + semantic_search 双证）。
