+++ crates/dasclaw_runtime/src/lib.rs
@@ -39,6 +39,7 @@ pub mod error;
 pub mod feature_flags;
 pub mod job;
 pub mod job_context;
+pub mod rate_limit;
 pub mod recording;
 pub mod secrets;
 pub mod tool;
@@ -47,5 +48,6 @@ pub use error::ToolError;
 pub use feature_flags::{SharedFeatureFlags, ToolFeatureFlags};
 pub use job::{JobState, StateTransition, TokenBudgetExceeded};
 pub use job_context::JobContextCore;
+pub use rate_limit::{LimitType, RateLimitError, RateLimitResult, RateLimiter};
 pub use recording::{HttpExchange, HttpExchangeRequest, HttpExchangeResponse, HttpInterceptor};
 pub use tool::Tool;
